Set a UUID based on the machine name when a UUID is unspecified.
Hyperkit uses this UUID to assign a MAC address that the driver is expecting. The driver searches for this MAC address in the DHCP lease file to resolve the machine.

Using a UUID based on machine name minimizes entries in the DHCP lease file. It also provides IP address consistency while the DHCP lease is valid.

Signed-off-by: John Coyle <dx9err@gmail.com>