Create uuid-to-mac-addr subcommand
The goal is to split out all functionality that needs to run as root
into subcommands that can then execute via the setuid logic. This will
then allow the driver itself to run under the regular user id.

This will make the code more secure, and also make sure e.g. the cache
directory will be owned by the user and not by root.

These internal commands are not scheduled via cobra to make sure privileges
can be dropped before cobra runs, which would otherwise run Pre command hooks
as root etc. It is a feature that internal commands will not be included in
the usage output.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>